+++ Pandora/Pattern/Functor/Applicative.hs
@@ -8,7 +8,7 @@ infixl 4 <*>, <*, *>
 
 {- |
 > When providing a new instance, you should ensure it satisfies the three laws:
-> * Composition: (.) <$> u <*> v <*> w ≡ u <*> (v <*> w)
+> * Interpreted: (.) <$> u <*> v <*> w ≡ u <*> (v <*> w)
 > * Left interchange: x <*> (f <$> y) ≡ (. f) <$> x <*> y
 > * Right interchange: f <$> (x <*> y) ≡ (f .) <$> x <*> y
 -}