+++ Pandora/Pattern/Functor/Invariant.hs
@@ -3,7 +3,7 @@ module Pandora.Pattern.Functor.Invariant (Invariant (..)) where
 {- |
 > When providing a new instance, you should ensure it satisfies the two laws:
 > Identity morphisms: invmap identity identity = identity
-> Composition of morphisms: invmap g j . invmap f h = invmap (g . f) (h . j)
+> Interpreted of morphisms: invmap g j . invmap f h = invmap (g . f) (h . j)
 -}
 
 class Invariant (t :: * -> *) where