+++ Pandora/Pattern/Junction.hs
@@ -2,4 +2,4 @@ module Pandora.Pattern.Junction (module Exports) where
 
 import Pandora.Pattern.Junction.Schemes as Exports
 import Pandora.Pattern.Junction.Transformer as Exports
-import Pandora.Pattern.Junction.Composition as Exports
+import Pandora.Pattern.Junction.Interpreted as Exports